+++ NEWS.md
@@ -31,6 +31,7 @@
 - Added tests of the common parameters of `make.`-prefixed functions
 - Documentation of save-as-you-go parameters in `make.`-prefixed functions has been updated, as it was outdated after an update to the underlying C library. The correct output file formats and file names are now described.
 - Update journal paper link and add a link to the alternate package guide (the "Templates" page of C documentation) in README and vignette.
+- Changed name of `startIndex` parameter of `change.label.to.values` to `skip` to avoid confusion with genomicSimulation internal indexes (`see.group.data(data.type="X")`), which are not valid inputs to this function. This is a non-breaking change, all existing scripts using `startIndex` still function exactly as before. 
 
 ## Bug Fixes